Add Wafel TTC RNG Bruteforcer + Data
Rewrote the pendulum bruteforcer using wafel which now runs an order of magnitude faster (6min instead of 6hrs).

Parameterize the data in terms of the frame that a squish cancel is possible.

The human element is still needed to TAS different RNG values to see if it can be completed (different yaw makes it harder, and RNG after the pendulum might be bad)